PEP8 and refactorization for reuse
